fix: change integer to long type in jpa repo generics

+++ src/main/java/com/biwaby/financialtracker/repository/CurrencyRepository.java
@@ -3,5 +3,5 @@ package com.biwaby.financialtracker.repository;
 import com.biwaby.financialtracker.entity.Currency;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CurrencyRepository extends JpaRepository<Currency, Integer> {
+public interface CurrencyRepository extends JpaRepository<Currency, Long> {
 }
