+++ src/main/java/com/biwaby/financialtracker/repository/SavingsAccountRepository.java
@@ -3,5 +3,5 @@ package com.biwaby.financialtracker.repository;
 import com.biwaby.financialtracker.entity.SavingsAccount;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SavingsAccountRepository extends JpaRepository<SavingsAccount, Integer> {
+public interface SavingsAccountRepository extends JpaRepository<SavingsAccount, Long> {
 }
