+++ src/main/java/com/biwaby/financialtracker/repository/SavingsTransactionRepository.java
@@ -3,5 +3,5 @@ package com.biwaby.financialtracker.repository;
 import com.biwaby.financialtracker.entity.SavingsTransaction;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SavingsTransactionRepository extends JpaRepository<SavingsTransaction, Integer> {
+public interface SavingsTransactionRepository extends JpaRepository<SavingsTransaction, Long> {
 }
