+++ src/main/java/com/biwaby/financialtracker/repository/WalletTransactionRepository.java
@@ -3,5 +3,5 @@ package com.biwaby.financialtracker.repository;
 import com.biwaby.financialtracker.entity.WalletTransaction;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface WalletTransactionRepository extends JpaRepository<WalletTransaction, Integer> {
+public interface WalletTransactionRepository extends JpaRepository<WalletTransaction, Long> {
 }
